Update _index.md
Hi Edward, I started an entry for you as a mentor in another folder. I'm going to delete that file and replace it with this one. I'm changing the `user_groups` in this commit so that you are moved into the mentors group. Other than that I think this looks great so I'll try to tidy up that other file and make the replacement.